src/requirements.txt
langchain
langchain-community==0.2.15
langchain-huggingface
langchain-experimental
pypdf2
pycryptodome
sentence-transformers==3.0.1
llama-cpp-python
neo4j==5.21.0
jproperties
pydantic
camelot-py
opencv-python
ghostscript
transformers
pdf2image
pytesseract
chromadb
rank_bm25
simsimd
